refactor(jruby): DriverFactory helpers as Ext (matches AuthTokens / BookmarkManagers / ...)
The previous JRuby `Internal::DriverFactory` Ruby subclass conflated
two concerns: providing Ruby helpers and giving testkit a class to
subclass for the `protected` resolver / clock overrides. Only the
overrides actually need a subclass.

- New `Ext::DriverFactory` module carries the helpers
  (`to_domain_name_resolver`, `to_clock`, the `new_instance`
  Ruby-Hash-config wrapper). Prepended onto Java's
  `org.neo4j.driver.internal.DriverFactory` the same way every other
  Ext module is wired (AuthTokens / BookmarkManagers /
  GraphDatabase / ...).
- `Neo4j::Driver::Internal::DriverFactory` becomes a plain constant
  alias to the Java class — same shape as the
  `StaticAuthTokenManager` alias right next to it. testkit
  subclasses that, naturally inheriting the prepended Ext methods,
  and overrides the two protected seams in pure Ruby.

MRI's `Internal::DriverFactory` Ruby class stays as the impl-side
sibling for testkit to subclass on that flavour.
