--- lib/jruby/neo4j/driver/internal/driver_factory.rb
@@ -1,58 +1,0 @@
-# frozen_string_literal: true
-
-module Neo4j
-  module Driver
-    module Internal
-      # JRuby flavour of the impl-agnostic `Internal::DriverFactory`
-      # seam. Subclasses Java's `org.neo4j.driver.internal.DriverFactory`
-      # so its `protected` hooks (`getDomainNameResolver`,
-      # `createClock`) can be overridden by a further Ruby subclass
-      # (testkit-backend's `DriverFactory`) without that subclass
-      # naming any Java type. Two converter helpers do the per-impl
-      # work:
-      #
-      #   to_domain_name_resolver(proc)   — wrap a Ruby `proc` returning
-      #                                     hostnames into Java's
-      #                                     DomainNameResolver SAM.
-      #   to_clock(clock)                 — wrap any Ruby `clock` with
-      #                                     `#now_millis` into a
-      #                                     `java.time.Clock`.
-      #
-      # `new_instance` adds a Ruby-friendly signature on top of Java's
-      # 4-arg form: callers hand `(uri, auth_token_manager, config)`
-      # — always a manager (callers wrap bare AuthTokens in
-      # `StaticAuthTokenManager`, same as Java's testkit-backend) — and
-      # we fill in the nil `ClientCertificateManager` plus the
-      # Ruby→Java config conversion.
-      class DriverFactory < Java::OrgNeo4jDriverInternal::DriverFactory
-        include Ext::ConfigConverter
-        include Ext::ExceptionCheckable
-
-        def to_domain_name_resolver(resolver_proc)
-          return nil unless resolver_proc
-
-          ->(name) do
-            resolver_proc.call(name).map do |addr|
-              java.net.InetAddress.get_by_name(addr)
-            rescue java.net.UnknownHostException => e
-              raise java.lang.RuntimeException.new(e)
-            end
-          end
-        end
-
-        def to_clock(clock)
-          Ext::Internal::ClockAdapter.new(clock)
-        end
-
-        def new_instance(uri, auth_token_manager, config = {})
-          check do
-            super(java.net.URI.create(uri.to_s),
-                  auth_token_manager,
-                  nil, # ClientCertificateManager — wired in a later slice
-                  to_java_config(Neo4j::Driver::Config, **config))
-          end
-        end
-      end
-    end
-  end
-end
